requirements.txt
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
channels==4.2.0
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
Django==5.1.6
django-choices-field==2.3.0
django-cors-headers==4.7.0
graphql-core==3.2.6
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
service-identity==24.2.0
setuptools==75.8.2
six==1.17.0
sqlparse==0.5.3
strawberry-graphql==0.261.1
strawberry-graphql-django==0.57.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
zope.interface==7.2
